Rename self-employed pension input to desired
