+++ changelog.d/fixed/8389.md
@@ -1,1 +1,1 @@
-Enforced the section 415(c) annual additions limit on self-employed pension contributions.
+Enforced the section 415(c) annual additions limit on desired self-employed pension contributions.
