+++ policyengine_us/parameters/gov/states/ut/tax/property/homeowner_renter_relief/nontaxable_income/sources.yaml
@@ -11,7 +11,7 @@ values:
     - capped_traditional_401k_contributions
     - capped_traditional_403b_contributions
     - capped_traditional_ira_contributions
-    - capped_self_employed_pension_contributions
+    - self_employed_pension_contributions
     - veterans_benefits
     - workers_compensation
 metadata:
