+++ policyengine_us/tests/policy/baseline/gov/states/mn/tax/property/homestead_credit_refund/mn_homestead_credit_refund.yaml
@@ -77,7 +77,7 @@
       person1:
         age: 40
         self_employment_income: 50_000
-        self_employed_pension_contributions: 10_000
+        self_employed_pension_contributions_desired: 10_000
         is_tax_unit_head: true
     tax_units:
       tax_unit:
