+++ policyengine_us/tests/policy/baseline/household/expense/retirement/self_employed_pension_contribution_limit.yaml
@@ -4,10 +4,10 @@
     people:
       person1:
         self_employment_income: 100_000
-        self_employed_pension_contributions: 100_000
+        self_employed_pension_contributions_desired: 100_000
   output:
     self_employed_pension_contribution_limit: 72_000
-    capped_self_employed_pension_contributions: 72_000
+    self_employed_pension_contributions: 72_000
     self_employed_pension_contribution_ald_person: 72_000
 
 - name: Self-employed pension contributions are also capped at self-employment compensation.
@@ -16,10 +16,10 @@
     people:
       person1:
         self_employment_income: 50_000
-        self_employed_pension_contributions: 100_000
+        self_employed_pension_contributions_desired: 100_000
   output:
     self_employed_pension_contribution_limit: 50_000
-    capped_self_employed_pension_contributions: 50_000
+    self_employed_pension_contributions: 50_000
     self_employed_pension_contribution_ald_person: 50_000
 
 - name: Unrelated employee elective deferrals do not reduce the self-employed pension cap.
@@ -31,12 +31,12 @@
         employment_income: 100_000
         self_employment_income: 100_000
         traditional_401k_contributions: 50_000
-        self_employed_pension_contributions: 72_000
+        self_employed_pension_contributions_desired: 72_000
   output:
     elective_deferral_limit: 35_750
     capped_traditional_401k_contributions: 35_750
     self_employed_pension_contribution_limit: 72_000
-    capped_self_employed_pension_contributions: 72_000
+    self_employed_pension_contributions: 72_000
 
 - name: Annual additions cap does not apply to IRA contributions.
   period: 2026
@@ -46,9 +46,9 @@
         age: 50
         self_employment_income: 100_000
         traditional_ira_contributions: 9_000
-        self_employed_pension_contributions: 72_000
+        self_employed_pension_contributions_desired: 72_000
   output:
     ira_contribution_limit: 8_000
     capped_traditional_ira_contributions: 8_000
     self_employed_pension_contribution_limit: 72_000
-    capped_self_employed_pension_contributions: 72_000
+    self_employed_pension_contributions: 72_000
