+++ policyengine_us/variables/household/expense/retirement/self_employed_pension_contributions.py
@@ -7,6 +7,13 @@ class self_employed_pension_contributions(Variable):
     label = "Self-employed pension contributions"
     unit = USD
     documentation = (
-        "Pension plan contributions associated with plans for the self employed."
+        "Self-employed pension contributions after applying the section "
+        "415(c) annual additions limit."
     )
     definition_period = YEAR
+    reference = "https://www.law.cornell.edu/uscode/text/26/415#c"
+
+    def formula(person, period, parameters):
+        desired = person("self_employed_pension_contributions_desired", period)
+        limit = person("self_employed_pension_contribution_limit", period)
+        return min_(desired, limit)
